+++ app/src/pages/hooks/_index.tsx
@@ -1,17 +1,17 @@
 import { Link } from 'react-router-dom';
-import { ComponentPage } from '../components/layout/ComponentPage';
-import { ExampleSection } from '../components/ui/ExampleSection';
+import { ComponentPage } from '../../components/layout/ComponentPage';
+import { ExampleSection } from '../../components/ui/ExampleSection';
 
 const hooks = [
   {
     name: 'useActionModal',
     path: '/hooks/useactionmodal',
-    description: 'Programmatically display alert and confirmation modals with async/await support.'
+    description: 'Programmatically display alert and confirmation modals with async/await support.',
   },
   {
     name: 'useToast',
     path: '/hooks/usetoast',
-    description: 'Display toast notifications with different types, actions, and customizable duration.'
+    description: 'Display toast notifications with different types, actions, and customizable duration.',
   },
 ];
 
@@ -21,10 +21,7 @@ export const HooksPage = () => {
       title='Hooks'
       description='Custom React hooks that provide additional functionality and state management for your components.'
     >
-      <ExampleSection 
-        title='Available Hooks'
-        description='Browse through our collection of custom React hooks.'
-      >
+      <ExampleSection title='Available Hooks' description='Browse through our collection of custom React hooks.'>
         <div className='grid grid-cols-1 md:grid-cols-2 gap-6'>
           {hooks.map((hook) => (
             <Link
@@ -35,25 +32,21 @@ export const HooksPage = () => {
               <h3 className='text-xl font-semibold text-white group-hover:text-primary transition-colors mb-2'>
                 {hook.name}
               </h3>
-              <p className='text-gray-400 text-sm'>
-                {hook.description}
-              </p>
-              <div className='mt-4 flex items-center text-primary text-sm'>
-                View Details →
-              </div>
+              <p className='text-gray-400 text-sm'>{hook.description}</p>
+              <div className='mt-4 flex items-center text-primary text-sm'>View Details →</div>
             </Link>
           ))}
         </div>
       </ExampleSection>
 
-      <ExampleSection 
+      <ExampleSection
         title='Provider Setup'
         description='Most hooks require their respective providers to be set up in your application.'
       >
         <div className='bg-gray-700 p-4 rounded-lg'>
           <h4 className='text-white font-semibold mb-2'>Complete Setup</h4>
           <pre className='text-sm text-gray-300 overflow-x-auto'>
-{`import { ActionModalProvider, ToastProvider } from '@moondreamsdev/dreamer-ui/providers';
+            {`import { ActionModalProvider, ToastProvider } from '@moondreamsdev/dreamer-ui/providers';
 
 function App() {
   return (
